+++ src/locale/ru/ru.js
@@ -145,7 +145,19 @@ export default {
     user: 'Пользователь | Пользователи',
     role: 'Роль | Роли',
     calendar: 'Календарь | Календари',
+    direction: 'Направление',
+    gateway: 'Шлюз | Шлюзы',
+    hangupCause: 'Причина окончания',
+    hasOption: 'Has option',
+    hasRecording: 'Запись разговора',
+    amdResult: 'AMD результат',
+    evaluation: 'Оценивание',
+    ratedBy: 'Оценено кем',
+    talkDuration: 'Длительность разговора',
+    totalDuration: 'Общая длительность',
+    transcription: 'Транскрипция',
     queue: {
+      queue: 'Очередь | Очереди',
       type: {
         [QueueType.INBOUND_QUEUE]: 'Входящая очередь',
         [QueueType.OFFLINE_QUEUE]: 'Оффлайн очередь',
@@ -159,6 +171,7 @@ export default {
       },
     },
     agent: {
+      agent: 'Оператор | Операторы',
       status: {
         [AgentStatus.Online]: 'Онлайн',
         [AgentStatus.Pause]: 'Пауза',
@@ -441,6 +454,90 @@ export default {
       title: 'Ошибка сохранения',
       label: 'Что-то пошло не так. Пожалуйста, попробуйте еще раз',
       exportToJson: 'Экспортировать в JSON',
+    },
+    filters: {
+      addFilter: ({ linked }) => {
+        return `${linked('reusable.add')} ${linked(
+          'reusable.filter',
+        ).toLowerCase()}`;
+      },
+      filterName: ({ linked }) => {
+        // because filter select has a 'column' label now
+        return linked('vocabulary.column');
+      },
+      filterValue: ({ linked }) => {
+        return linked('vocabulary.values');
+      },
+      filterValueFrom: ({ linked }) => {
+        const from = linked('reusable.from').toLowerCase();
+        return `${linked('vocabulary.values')} ${from}`;
+      },
+      filterLabel: ({ linked }) => {
+        return linked('vocabulary.labels');
+      },
+      agent: ({ linked }) => {
+        return linked('objects.agent.agent');
+      },
+      amdResult: ({ linked }) => {
+        return linked('objects.amdResult');
+      },
+      contact: ({ linked }) => {
+        return linked('vocabulary.contact');
+      },
+      createdAtFrom: ({ linked }) => {
+        return linked('reusable.from');
+      },
+      createdAtTo: ({ linked }) => {
+        return linked('reusable.to');
+      },
+      direction: ({ linked }) => {
+        return linked('objects.direction');
+      },
+      rated: ({ linked }) => {
+        return linked('objects.evaluation');
+      },
+      gateway: ({ linked }) => {
+        return linked('objects.gateway');
+      },
+      grantee: ({ linked }) => {
+        return linked('objects.grantee');
+      },
+      cause: ({ linked }) => {
+        return linked('objects.hangupCause');
+      },
+      hasOption: ({ linked }) => {
+        return linked('objects.hasOption'); ////
+      },
+      queue: ({ linked }) => {
+        return linked('objects.queue.queue');
+      },
+      hasFile: ({ linked }) => {
+        return linked('objects.hasRecording');
+      },
+      score: ({ linked }) => {
+        return linked('webitelUI.auditForm.score');
+      },
+      tags: ({ linked }) => {
+        return linked('vocabulary.tag');
+      },
+      talkSec: ({ linked }) => {
+        return linked('objects.talkDuration');
+      },
+      team: ({ linked }) => {
+        return linked('objects.team');
+      },
+      duration: ({ linked }) => {
+        return linked('objects.totalDuration');
+      },
+      hasTranscription: ({ linked }) => {
+        return linked('objects.transcription');
+      },
+      user: ({ linked }) => {
+        return linked('objects.user');
+      },
+      variable: ({ linked }) => {
+        return linked('vocabulary.variables');
+      },
     },
   },
   errorNotifications: {