+++ src/modules/Filters/v2/filters/components/preview/dynamic-filter-preview-info.vue
@@ -1,38 +1,19 @@
 <template>
   <section class="dynamic-filter-preview-info">
-    <header class="dynamic-filter-preview-info-header">
-      {{ props.name }}
+    <header class="dynamic-filter-preview-info__header">
+      <slot name="header"/>
     </header>
 
     <wt-divider />
 
-    <!--    TODO: how to visually represent id/value types? -->
-    <ul v-if="isArrayValue">
-      <li
-        v-for="(el, index) of value"
-        :key="index"
-      >
-        {{ el }}
-      </li>
-    </ul>
+    <div class="dynamic-filter-preview-info__body">
+      <slot name="default" />
+    </div>
 
-    <article v-else>
-      {{ props.value }}
-    </article>
   </section>
 </template>
 
 <script lang="ts" setup>
-import type { FilterName, FilterValue } from '../../types/Filter';
-
-interface Props {
-  name: FilterName;
-  value: FilterValue;
-}
-
-const props = defineProps<Props>();
-
-const isArrayValue = Array.isArray(props.value);
 </script>
 
 <style lang="scss" scoped>
@@ -40,9 +21,15 @@ const isArrayValue = Array.isArray(props.value);
   display: flex;
   flex-direction: column;
   gap: var(--spacing-xs);
-}
 
-.dynamic-filter-preview-info-header {
-  @extend %typo-subtitle-1;
+  &__header {
+    @extend %typo-subtitle-1;
+  }
+
+  &__body {
+    display: flex;
+    flex-direction: column;
+    gap: var(--spacing-xs);
+  }
 }
 </style>