+++ src/modules/Filters/v2/filters/components/preview/dynamic-filter-preview.vue
@@ -4,7 +4,7 @@
       <wt-tooltip :disabled="configFormVisible">
         <template #activator>
           <wt-chip color="primary">
-            {{ props.filter.label || props.filter.name }}
+            {{ props.filter.label || t(`webitelUI.filters.${props.filter.name}`) }}
             <wt-icon-btn
               icon="close--filled"
               size="sm"
@@ -14,10 +14,16 @@
         </template>
 
         <template #default>
-          <dynamic-filter-preview-info
-            :name="props.filter.name"
-            :value="props.filter.value"
-          />
+          <dynamic-filter-preview-info>
+            <template #header>
+              {{ t(`webitelUI.filters.${props.filter.name}`) }}
+            </template>
+
+            <template>
+              <slot name="info" />
+            </template>
+          </dynamic-filter-preview-info>
+
         </template>
       </wt-tooltip>
     </template>
@@ -34,11 +40,14 @@ import WtIconBtn from '../../../../../../components/wt-icon-btn/wt-icon-btn.vue'
 import WtTooltip from '../../../../../../components/wt-tooltip/wt-tooltip.vue';
 import type { IFilter } from '../../types/Filter';
 import DynamicFilterConfigView from '../config/dynamic-filter-config-view.vue';
-import DynamicFilterPreviewInfo from './dynamic-filter-preview-info.vue';
+import DynamicFilterPreviewInfo from "./dynamic-filter-preview-info.vue";
+import {useI18n} from "vue-i18n";
 
 interface Props {
   filter: IFilter;
 }
+
+const { t } = useI18n();
 
 const props = defineProps<Props>();