+++ src/modules/Filters/v2/filters/components/values/agent/config.js
@@ -1,0 +1,4 @@
+import AgentsAPI from '../../../../../../../api/clients/agents/agents.js';
+
+export const searchMethod = AgentsAPI.getLookup;
+export const localePath = '';