+++ src/modules/Filters/v2/filters/components/values/amd-result/amd-result-filter-value-preview.vue
@@ -1,0 +1,20 @@
+<template>
+  <ul>
+    <li
+      v-for="(result, index) of props.value"
+      :key="index"
+    >
+      {{ result }}
+    </li>
+  </ul>
+</template>
+
+<script lang="ts" setup>
+const props = defineProps<{
+  value: number[];
+}>();
+
+
+</script>
+
+<style lang="scss" scoped></style>