+++ src/modules/Filters/v2/filters/components/values/created-at-to/created-at-to-filter-value-field.vue
@@ -1,0 +1,23 @@
+<template>
+  <wt-datepicker
+    v-if="model"
+    :label="`${t('webitelUI.filters.createdAtTo')}:`"
+    :value="model"
+    @input="model = $event"
+  />
+</template>
+
+<script lang="ts" setup>
+import { endOfToday } from 'date-fns';
+import { useI18n } from 'vue-i18n';
+
+const model = defineModel<number>();
+
+const { t } = useI18n();
+
+if (!model.value) {
+  model.value = endOfToday().getTime();
+}
+</script>
+
+<style lang="scss" scoped></style>