+++ src/modules/Filters/v2/filters/components/values/created-at-to/created-at-to-filter-value-preview.vue
@@ -1,0 +1,15 @@
+<template>
+  <div>{{date}}</div>
+</template>
+
+<script lang="ts" setup>
+import { ref } from 'vue';
+
+const props = defineProps<{
+  value: number[];
+}>();
+
+const date = ref(new Date(+props.value).toLocaleDateString());
+</script>
+
+<style lang="scss" scoped></style>