+++ src/modules/Filters/v2/filters/components/values/index.js
@@ -1,0 +1,87 @@
+import AgentFilter from './agent/agent-filter-value-field.vue';
+import AgentFilterPreview from './agent/agent-filter-value-preview.vue';
+import AmdResultFilter from './amd-result/amd-result-filter-value-field.vue';
+import AmdResultFilterPreview from './amd-result/amd-result-filter-value-preview.vue';
+import CauseFilter from './cause/cause-filter-value-field.vue';
+import CauseFilterPreview from './cause/cause-filter-value-preview.vue';
+import ContactFilter from './contact/contact-filter-value-field.vue';
+import ContactFilterPreview from './contact/contact-filter-value-preview.vue';
+import CreatedAtFromFilter from './created-at-from/created-at-from-filter-value-field.vue';
+import CreatedAtFromFilterPreview from './created-at-from/created-at-from-filter-value-preview.vue';
+import CreatedAtToFilter from './created-at-to/created-at-to-filter-value-field.vue';
+import CreatedAtToFilterPreview from './created-at-to/created-at-to-filter-value-preview.vue';
+import DirectionFilter from './direction/direction-filter-value-field.vue';
+import DirectionFilterPreview from './direction/direction-filter-value-preview.vue';
+import GatewayFilter from './gateway/gateway-filter-value-field.vue';
+import GatewayFilterPreview from './gateway/gateway-filter-value-preview.vue';
+import GranteeFilter from './grantee/grantee-filter-value-field.vue';
+import GranteeFilterPreview from './grantee/grantee-filter-value-preview.vue';
+import HasFileFilter from './has-file/has-file-filter-value-field.vue';
+import HasFileFilterPreview from './has-file/has-file-filter-value-preview.vue';
+import HasTranscriptionFilter from './has-transcription/has-transcription-filter-value-field.vue';
+import HasTranscriptionFilterPreview from './has-transcription/has-transcription-filter-value-preview.vue';
+import QueueFilter from './queue/queue-filter-value-field.vue';
+import QueueFilterPreview from './queue/queue-filter-value-preview.vue';
+import RatedFilter from './rated/rated-filter-value-field.vue';
+import RatedFilterPreview from './rated/rated-filter-value-preview.vue';
+import RatedByFilter from './rated-by/rated-by-filter-value-field.vue';
+import RatedByFilterPreview from './rated-by/rated-by-filter-value-preview.vue';
+import ScoreFilter from './score/score-from-to-filter-value-field.vue';
+import ScoreFilterPreview from './score/score-from-to-filter-value-preview.vue';
+import TagFilter from './tag/tag-filter-value-field.vue';
+import TagFilterPreview from './tag/tag-filter-value-preview.vue';
+import TalkDuration from './talk-duration/talk-duration-filter-value-field.vue';
+import TalkDurationPreview from './talk-duration/talk-duration-filter-value-preview.vue';
+import TeamFilter from './team/team-filter-value-field.vue';
+import TeamFilterPreview from './team/team-filter-value-preview.vue';
+import TotalDurationFilter from './total-duration/total-duration-filter-value-field.vue';
+import TotalDurationFilterPreview from './total-duration/total-duration-filter-value-preview.vue';
+import UserFilter from './user/user-filter-value-field.vue';
+import UserFilterPreview from './user/user-filter-value-preview.vue';
+import VariableFilter from './variable/variable-filter-value-field.vue';
+import VariableFilterPreview from './variable/variable-filter-value-preview.vue';
+
+export {
+  AgentFilter,
+  AgentFilterPreview,
+  AmdResultFilter,
+  AmdResultFilterPreview,
+  CauseFilter,
+  CauseFilterPreview,
+  ContactFilter,
+  ContactFilterPreview,
+  CreatedAtFromFilter,
+  CreatedAtFromFilterPreview,
+  CreatedAtToFilter,
+  CreatedAtToFilterPreview,
+  DirectionFilter,
+  DirectionFilterPreview,
+  GatewayFilter,
+  GatewayFilterPreview,
+  GranteeFilter,
+  GranteeFilterPreview,
+  HasFileFilter,
+  HasFileFilterPreview,
+  HasTranscriptionFilter,
+  HasTranscriptionFilterPreview,
+  QueueFilter,
+  QueueFilterPreview,
+  RatedByFilter,
+  RatedByFilterPreview,
+  RatedFilter,
+  RatedFilterPreview,
+  ScoreFilter,
+  ScoreFilterPreview,
+  TagFilter,
+  TagFilterPreview,
+  TalkDuration,
+  TalkDurationPreview,
+  TeamFilter,
+  TeamFilterPreview,
+  TotalDurationFilter,
+  TotalDurationFilterPreview,
+  UserFilter,
+  UserFilterPreview,
+  VariableFilter,
+  VariableFilterPreview,
+};