+++ src/modules/Filters/v2/filters/components/values/team/config.js
@@ -1,0 +1,4 @@
+import TeamsAPI from '../../../../../../../api/clients/teams/teams.js';
+
+export const searchMethod = TeamsAPI.getLookup;
+export const localePath = '';