+++ src/modules/Filters/v2/filters/components/values/users/config.js
@@ -1,4 +1,0 @@
-// import { UsersApi}
-
-export const searchMethod = UsersAPI.getLookup;
-export const localePath = ''; /* label locale goes here */